// frontend/src/services/openapi/models/ChapterMetadataResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ChapterMetadataResponse = {
    id: string;
    title?: string | null;
    chapterNumber: number;
    volumeNumber?: number | null;
    createdAt: string;
    updatedAt: string;
};